@@ -1,97 +1,107 @@
 {
-	"name": "peopledesk",
-	"version": "1.1.0",
-	"private": true,
-	"main": "index.js",
-	"dependencies": {
-		"@ant-design/charts": "^1.3.6",
-		"@ant-design/plots": "^1.0.9",
-		"@emotion/react": "^11.4.1",
-		"@emotion/styled": "^11.3.0",
-		"@fvilers/disable-react-devtools": "^1.3.0",
-		"@microsoft/signalr": "^6.0.1",
-		"@mui/core": "^5.0.0-alpha.54",
-		"@mui/icons-material": "^5.0.0",
-		"@mui/lab": "^5.0.0-alpha.55",
-		"@mui/material": "^5.0.0",
-		"@mui/styles": "^5.0.0",
-		"@mui/x-data-grid": "^5.12.3",
-		"@mui/x-date-pickers": "^5.0.13",
-		"@reduxjs/toolkit": "^1.6.0",
-		"antd": "^4.22.4",
-		"apexcharts": "^3.29.0",
-		"axios": "0.21.1",
-		"bootstrap": "^4.5.2",
-		"crypto-js": "^4.1.1",
-		"date-fns": "^2.26.0",
-		"exceljs": "^4.3.0",
-		"file-saver": "^2.0.5",
-		"formik": "^2.2.9",
-		"lodash": "^4.17.21",
-		"material-ui-dropzone": "^3.5.0",
-		"materialui-daterange-picker": "^1.1.92",
-		"mgr-pdf-viewer-react": "^1.0.3",
-		"moment": "^2.29.1",
-		"mui-datatables": "^4.2.2",
-		"peerjs": "^1.3.2",
-		"rc-year-calendar": "^1.0.2",
-		"react": "^17.0.2",
-		"react-apexcharts": "^1.3.9",
-		"react-beautiful-dnd": "^13.1.1",
-		"react-bootstrap": "^1.6.1",
-		"react-confirm-alert": "^2.7.0",
-		"react-dom": "^17.0.2",
-		"react-grid-layout": "^1.3.4",
-		"react-icons": "^4.9.0",
-		"react-organizational-chart": "2.2.0",
-		"react-quill": "^1.3.5",
-		"react-redux": "^7.2.4",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "^5.0.1",
-		"react-select": "^5.0.0",
-		"react-to-print": "^2.14.4",
-		"react-toastify": "^7.0.4",
-		"redux": "^4.1.0",
-		"redux-persist": "^6.0.0",
-		"swiper": "6.8.2",
-		"yarn": "^1.22.19",
-		"yup": "^0.32.9"
-	},
-	"scripts": {
-		"start": "set PORT=3013 && react-scripts --max_old_space_size=8048 start -o",
-		"build": "GENERATE_SOURCEMAP=false react-scripts --max_old_space_size=16096 build",
-		"build-dev": "react-scripts build",
-		"test": "cross-env CI=true react-scripts test --env=jsdom",
-		"eject": "react-scripts eject",
-		"lint": "eslint ./src/"
-	},
-	"eslintConfig": {
-		"extends": [
-			"react-app"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"@babel/core": "^7.16.7",
-		"@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.5",
-		"@babel/plugin-transform-object-assign": "^7.16.7",
-		"@babel/preset-env": "^7.16.7",
-		"babel-loader": "^8.2.3",
-		"babel-preset-react": "^6.24.1",
-		"socket.io-client": "^4.4.0",
-		"webpack": "^4.44.2",
-		"webpack-cli": "^4.9.1",
-		"webpack-remove-debug": "^0.1.0"
-	}
+  "name": "peopledesk",
+  "version": "1.1.0",
+  "private": true,
+  "main": "index.js",
+  "dependencies": {
+    "@ant-design/charts": "^1.3.6",
+    "@ant-design/plots": "^1.0.9",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@fvilers/disable-react-devtools": "^1.3.0",
+    "@microsoft/signalr": "^6.0.1",
+    "@mui/icons-material": "^5.0.0",
+    "@mui/lab": "^5.0.0-alpha.152",
+    "@mui/material": "^5.0.0",
+    "@mui/styles": "^5.0.0",
+    "@mui/x-data-grid": "^5.0.0",
+    "@mui/x-date-pickers": "^5.0.0",
+    "@reduxjs/toolkit": "^1.6.0",
+    "antd": "^4.22.4",
+    "apexcharts": "^3.29.0",
+    "axios": "0.21.1",
+    "bootstrap": "^4.5.2",
+    "crypto-js": "^4.2.0",
+    "date-fns": "^2.26.0",
+    "exceljs": "^4.3.0",
+    "file-saver": "^2.0.5",
+    "formik": "^2.2.9",
+    "lodash": "^4.17.21",
+    "material-ui-dropzone": "^3.5.0",
+    "materialui-daterange-picker": "^1.1.92",
+    "mgr-pdf-viewer-react": "^1.0.3",
+    "moment": "^2.29.1",
+    "mui-datatables": "^4.2.2",
+    "peerjs": "^1.3.2",
+    "rc-year-calendar": "^1.0.2",
+    "react": "18.2.0",
+    "react-apexcharts": "^1.3.9",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-bootstrap": "^1.6.1",
+    "react-confirm-alert": "^2.7.0",
+    "react-dom": "18.2.0",
+    "react-grid-layout": "^1.3.4",
+    "react-icons": "^4.9.0",
+    "react-organizational-chart": "2.2.0",
+    "react-quill": "^1.3.5",
+    "react-redux": "^7.2.4",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.0.0",
+    "react-to-print": "^2.14.4",
+    "react-toastify": "^7.0.4",
+    "redux": "^4.1.0",
+    "redux-persist": "^6.0.0",
+    "sass": "^1.69.5",
+    "swiper": "6.8.2",
+    "yarn": "^1.22.19",
+    "yup": "^0.32.9"
+  },
+  "devDependencies": {
+    "@ant-design/icons": "^5.2.6",
+    "@babel/core": "^7.16.7",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.5",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-object-assign": "^7.16.7",
+    "@babel/preset-env": "^7.16.7",
+    "@types/estree": "^1.0.5",
+    "@types/jest": "^27.0.4",
+    "@types/node": "^16.11.18",
+    "@types/react": "^18.0.2",
+    "@types/react-dom": "^18.0.2",
+    "@types/react-router-dom": "^5.3.3",
+    "babel-loader": "^8.2.3",
+    "babel-runtime": "^6.26.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "socket.io-client": "^4.4.0",
+    "typescript": "^4.5.5",
+    "webpack": "^4.44.2",
+    "webpack-cli": "^4.9.1",
+    "webpack-remove-debug": "^0.1.0"
+  },
+  "scripts": {
+    "start": "set PORT=3013 && set NODE_OPTIONS=--max_old_space_size=6144 && react-scripts start",
+    "build": "GENERATE_SOURCEMAP=false react-scripts --max_old_space_size=16096 build",
+    "build-dev": "react-scripts build",
+    "test": "cross-env CI=true react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "lint": "eslint ./src/"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
