@@ -1,24 +1,24 @@
 /* eslint-disable eqeqeq */
 /* eslint-disable react-hooks/exhaustive-deps */
-import { Avatar } from "@material-ui/core";
+import "../style.css";
+import { TablePagination } from "@mui/material";
 import { AddOutlined, ModeEditOutlineOutlined } from "@mui/icons-material";
 import { Tooltip } from "@mui/material";
 import { Form, Formik } from "formik";
 import React, { useEffect, useRef, useState } from "react";
 import { shallowEqual, useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
-import AntTable from "../../../../common/AntTable";
-import { getPeopleDeskAllLanding } from "../../../../common/api";
 import FormikSelect from "../../../../common/FormikSelect";
 import Loading from "../../../../common/loading/Loading";
 import NotPermittedPage from "../../../../common/notPermitted/NotPermittedPage";
 import PrimaryButton from "../../../../common/PrimaryButton";
-import ViewModal from "../../../../common/ViewModal";
 import { setFirstLevelNameAction } from "../../../../commonRedux/reduxForLocalStorage/actions";
 import { customStyles } from "../../../../utility/selectCustomStyle";
 import { yearDDLAction } from "../../../../utility/yearDDL";
-import "../style.css";
-import CreateYearlyPolicyModal from "./CreateYearlyPolicyModal";
+import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
+import { getYearlyPolicyLanding } from "./helper";
+import ScrollableTable from "../../../../common/ScrollableTable";
+import NoResult from "common/NoResult";
 
 let date = new Date();
 let currentYear = date.getFullYear();
@@ -28,18 +28,21 @@ const initData = {
 };
 
 const YearlyLeavePolicy = () => {
-  const [show, setShow] = useState(false);
+  const [allPolicy, setAllPolicy] = useState([]);
   const [landingData, setLandingData] = useState([]);
-  const [allData, setAllData] = useState([]);
-  const [singleData, setSingleData] = useState({});
   const [sortType, setSortType] = useState("desc");
-
+  const history = useHistory();
+  const [pages, setPages] = useState({
+    currentPage: 1,
+    pageSize: 25,
+    totalCount: 0,
+  });
   const [, setYear] = useState(null);
 
   const saveHandler = (values, cb) => {};
   const [loading, setLoading] = useState(false);
 
-  const { orgId, buId, wgId, wId } = useSelector(
+  const { orgId, buId, wgId } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -48,21 +51,6 @@ const YearlyLeavePolicy = () => {
     dispatch(setFirstLevelNameAction("Administration"));
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  const getData = (year) => {
-    getPeopleDeskAllLanding(
-      "EmploymentTypeWiseLeaveBalance",
-      orgId,
-      buId,
-      0,
-      setLandingData,
-      setAllData,
-      setLoading,
-      "",
-      year ? year : formikRef?.current?.values?.year?.value,
-      `${wgId}&strWorkplace=${wId}`
-    );
-  };
 
   // const filterData = (year) => {
   //   let data = [];
@@ -87,123 +75,54 @@ const YearlyLeavePolicy = () => {
     });
     setLandingData(newData);
   };
-
+  const getLanding = (page = pages, year = currentYear) => {
+    getYearlyPolicyLanding(
+      `/SaasMasterData/AllLeavePolicyLanding?businessUnitId=${buId}&PageNo=${page?.currentPage}&PageSize=${page?.pageSize}&IsForXl=false&intYear=${year}`,
+      setAllPolicy,
+      setPages,
+      setLoading
+    );
+  };
   useEffect(() => {
-    getData();
+    getLanding(pages, currentYear);
   }, [orgId, buId, wgId]);
 
   const { permissionList } = useSelector((state) => state?.auth, shallowEqual);
-
   let permission = null;
   permissionList.forEach((item) => {
     if (item?.menuReferenceId === 38) {
       permission = item;
     }
   });
+  // handleChangePage
+  const handleChangePage = (event, newPage) => {
+    setPages((prev) => {
+      return { ...prev, currentPage: newPage + 1 };
+    });
 
-  const columns = [
-    {
-      title: "SL",
-      render: (text, record, index) => index + 1,
-      sorter: false,
-      filter: false,
-      width: 20,
-    },
-    {
-      title: "Employment Type",
-      dataIndex: "EmploymentTypeName",
-      sorter: true,
-      filter: false,
-    },
-    {
-      title: "Leave Type",
-      dataIndex: "LeaveTypeName",
-      sorter: true,
-      filter: false,
-    },
-    {
-      title: "Workplace Group",
-      dataIndex: "strWorkplaceGroup",
-      sorter: true,
-      filter: false,
-    },
-    {
-      title: "Workplace",
-      dataIndex: "strWorkplace",
-      sorter: true,
-      filter: false,
-    },
-    {
-      title: "Days",
-      dataIndex: "AllocatedLeave",
-      sorter: true,
-      filter: false,
-    },
-    {
-      title: "",
-      dataIndex: "",
-      render: (_, item) => (
-        <div
-          onClick={(e) => {
-            if (!permission?.isEdit)
-              return toast.warn("You don't have permission");
-            e.stopPropagation();
-            setShow(true);
-            setSingleData({
-              businessUnit: [
-                {
-                  value: item?.BusinessUnitId,
-                  label: item?.strBusinessUnit,
-                },
-              ],
-              workplaceGroup: [
-                {
-                  value: item?.intWorkplaceGroupId,
-                  label: item?.strWorkplaceGroup,
-                },
-              ],
-              workplace: [
-                {
-                  value: item?.intWorkplaceId,
-                  label: item?.strWorkplace,
-                },
-              ],
-              year: {
-                value: item?.YearId,
-                label: item?.YearId,
-              },
-              employmentType: {
-                value: item?.EmploymentTypeId,
-                label: item?.EmploymentTypeName,
-              },
-              leaveType: {
-                value: item?.LeaveTypeId,
-                label: item?.LeaveTypeName,
-              },
-              gender: {
-                value: item?.GenderId,
-                label: item?.GenderName,
-              },
-              days: item?.AllocatedLeave,
-              autoId: item?.AutoId,
-            });
-          }}
-        >
-          <Tooltip title="Edit" arrow>
-            <Avatar className="edit-icon-btn">
-              <ModeEditOutlineOutlined
-                sx={{
-                  color: "rgba(0, 0, 0, 0.6)",
-                }}
-              />
-            </Avatar>
-          </Tooltip>
-        </div>
-      ),
-      sorter: false,
-      filter: false,
-    },
-  ];
+    getLanding(
+      {
+        currentPage: newPage + 1,
+        pageSize: pages?.pageSize,
+        totalCount: pages?.totalCount,
+      },
+      currentYear
+    );
+  };
+
+  const handleChangeRowsPerPage = (event) => {
+    setPages((prev) => {
+      return { ...prev, pageSize: +event.target.value };
+    });
+    getLanding(
+      {
+        currentPage: pages?.currentPage,
+        pageSize: +event.target.value,
+        totalCount: pages?.totalCount,
+      },
+      currentYear
+    );
+  };
   const formikRef = useRef();
   return (
     <>
@@ -243,10 +162,10 @@ const YearlyLeavePolicy = () => {
                           isClearable={false}
                           onChange={(valueOption) => {
                             if (valueOption?.value) {
-                              getData(valueOption?.value);
+                              getLanding(pages, valueOption?.value);
                               // filterData(valueOption?.value);
                             } else {
-                              setLandingData({ Result: allData });
+                              getLanding(pages);
                             }
                             setFieldValue("year", valueOption);
                             setYear(valueOption?.value);
@@ -274,8 +193,10 @@ const YearlyLeavePolicy = () => {
                           onClick={(e) => {
                             if (!permission?.isCreate)
                               return toast.warn("You don't have permission");
-                            setSingleData({});
-                            setShow(true);
+                            // setShow(true);
+                            history.push(
+                              "/administration/leaveandmovement/yearlyLeavePolicy/create"
+                            );
                           }}
                           style={{ marginLeft: "16px" }}
                         />
@@ -293,9 +214,127 @@ const YearlyLeavePolicy = () => {
                       >
                         Year {values?.year ? values?.year.label : currentYear}
                       </h5>
-                      <div className="table-card-styled tableOne">
-                        <AntTable data={landingData} columnsData={columns} />
-                      </div>
+                      {Object.entries(allPolicy).length > 0 ? (
+                        <div className="table-card-styled tableOne">
+                          <ScrollableTable
+                            classes="salary-process-table"
+                            secondClasses="table-card-styled tableOne scroll-table-height"
+                            customClass="salary-details-custom"
+                          >
+                            <thead>
+                              <tr>
+                                {/* <th rowSpan="2" className="text-center">
+                                SL
+                              </th> */}
+                                <th
+                                  rowSpan="2"
+                                  style={{ minWidth: "100px" }}
+                                  className="text-center"
+                                >
+                                  Policy Name
+                                </th>
+                                <th rowSpan="2" className="text-center">
+                                  Leave Type
+                                </th>
+                                <th rowSpan="2" className="text-center">
+                                  Employee Type
+                                </th>
+                                <th rowSpan="2" className="text-center">
+                                  Gender{" "}
+                                </th>
+                                <th rowSpan="2" className="text-center">
+                                  Hr Position
+                                </th>
+                              </tr>
+                            </thead>
+                            <tbody>
+                              {Object.entries(allPolicy).map(
+                                ([workplaceName, policies]) => (
+                                  <React.Fragment key={workplaceName}>
+                                    {/* Row with workplace name */}
+                                    <tr>
+                                      <td colSpan="6">
+                                        {" "}
+                                        <span
+                                          style={{
+                                            fontWeight: "700",
+                                            color: "black",
+                                          }}
+                                        >
+                                          {workplaceName}
+                                        </span>
+                                      </td>
+                                    </tr>
+                                    {/* Rows with policy details */}
+                                    {policies.map((policy, index) => (
+                                      <tr key={index}>
+                                        <td>{policy?.strPolicyName}</td>
+                                        <td>{policy?.strLeaveTypeName}</td>
+                                        <td>
+                                          {policy?.employmentTypeList
+                                            ?.map(
+                                              (employmentType) =>
+                                                employmentType?.strEmploymentTypeName
+                                            )
+                                            ?.join(", ")}
+                                        </td>
+                                        <td>
+                                          {" "}
+                                          {policy?.genderListDTO
+                                            ?.map(
+                                              (gender) =>
+                                                gender?.strGenderName + " "
+                                            )
+                                            ?.join(", ")}
+                                        </td>
+                                        <td>
+                                          {" "}
+                                          {policy?.hrPositionListDTO
+                                            ?.map(
+                                              (pos) =>
+                                                pos?.strHrPositionName + " "
+                                            )
+                                            ?.join(", ")}
+                                        </td>
+                                        {orgId !== 0 && (
+                                          <td>
+                                            <Tooltip title="Edit" arrow>
+                                              <button
+                                                className="iconButton"
+                                                type="button"
+                                              >
+                                                <ModeEditOutlineOutlined
+                                                  onClick={() =>
+                                                    history.push(
+                                                      `/administration/leaveandmovement/yearlyLeavePolicy/edit/${policy?.policyId}`
+                                                    )
+                                                  }
+                                                  style={{ fontSize: "15px" }}
+                                                />
+                                              </button>
+                                            </Tooltip>
+                                          </td>
+                                        )}
+                                      </tr>
+                                    ))}
+                                  </React.Fragment>
+                                )
+                              )}
+                            </tbody>
+                          </ScrollableTable>
+                          <TablePagination
+                            rowsPerPageOptions={[5, 10, 15, 25, 100]}
+                            component="div"
+                            count={pages?.totalCount}
+                            rowsPerPage={pages?.pageSize}
+                            page={pages?.currentPage - 1}
+                            onPageChange={handleChangePage}
+                            onRowsPerPageChange={handleChangeRowsPerPage}
+                          />
+                        </div>
+                      ) : (
+                        <NoResult />
+                      )}
                     </div>
                   </div>
                 </div>
@@ -308,20 +347,6 @@ const YearlyLeavePolicy = () => {
           </>
         )}
       </Formik>
-      <ViewModal
-        size="lg"
-        title="Create yearly leave policy"
-        backdrop="static"
-        classes="default-modal preview-modal"
-        show={show}
-        onHide={() => setShow(false)}
-      >
-        <CreateYearlyPolicyModal
-          getData={getData}
-          singleData={singleData}
-          setShow={setShow}
-        />
-      </ViewModal>
     </>
   );
 };
