@@ -47,7 +47,7 @@ function BulkAddEditForm() {
   const { isCreate, isView } = location?.state?.state;
 
   //redux data
-  const { orgId, buId, employeeId, wgId } = useSelector(
+  const { orgId, buId, employeeId, wgId,wId } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -342,6 +342,7 @@ function BulkAddEditForm() {
                           touched={touched}
                           orgId={orgId}
                           buId={buId}
+                          wId={wId}
                           setLoading={setLoading}
                           setAllowanceAndDeductionDDL={
                             setAllowanceAndDeductionDDL
