@@ -28,11 +28,13 @@ export const createSalaryGenerateRequest = async (payload, setLoading, cb) => {
 
 // salary generate landing
 const currentYear = new Date().getFullYear();
+
 export const getSalaryGenerateRequestLanding = async (
   partName,
   orgId,
   buId,
   wgId,
+  wId,
   monthId,
   yearId,
   fromDate,
@@ -64,7 +66,7 @@ export const getSalaryGenerateRequestLanding = async (
         monthId || +currentMonth()
       }&intYearId=${
         yearId || currentYear
-      }&intWorkplaceGroupId=${wgId}&intBankOrWalletType=0${fromDateParams}${toDateParams}${wingParams}${soleDepoParams}${regionParams}${areaParams}${territoryParams}`
+      }&intWorkplaceGroupId=${wgId}&intWorkplaceId=${wId || 0}&intBankOrWalletType=0${fromDateParams}${toDateParams}${wingParams}${soleDepoParams}${regionParams}${areaParams}${territoryParams}`
     );
     if (res?.data) {
       const modifyRowData = res?.data?.map((itm) => {
