@@ -42,7 +42,7 @@ const SalaryGenerateCreate = () => {
   const dispatch = useDispatch();
 
   // redux
-  const { orgId, buId, employeeId, wgId, buName, wgName } = useSelector(
+  const { orgId, buId, employeeId, wgId, buName, wgName, wId, wName  } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -88,6 +88,7 @@ const SalaryGenerateCreate = () => {
       orgId,
       buId,
       wgId,
+      wId,
       setRowDto,
       setAllData,
       setLoading
@@ -195,6 +196,8 @@ const SalaryGenerateCreate = () => {
       strBusinessUnit: buName,
       intWorkplaceGroupId: wgId,
       strWorkplaceGroup: wgName,
+      intWorkplaceId: wId,
+      strWorkplace: wName,
       intWingId: values?.wing?.value || 0,
       intSoleDepoId: values?.soleDepo?.value || 0,
       intRegionId: values?.region?.value || 0,
@@ -879,6 +882,7 @@ const SalaryGenerateCreate = () => {
                               orgId,
                               buId,
                               wgId,
+                              wId,
                               values?.monthId,
                               values?.yearId,
                               values?.fromDate,
@@ -945,6 +949,7 @@ const SalaryGenerateCreate = () => {
                               orgId,
                               buId,
                               wgId,
+                              wId,
                               values?.monthId,
                               values?.yearId,
                               values?.fromDate,
