@@ -15,7 +15,10 @@ import {
   gray700,
 } from "../../../utility/customColor";
 import SingleNotice from "../../dashboard/dashboardModule/SingleNotice";
-import { getEmployeeDashboard } from "../../dashboard/helper";
+import {
+  getBirthAnniversary,
+  getEmployeeDashboard,
+} from "../../dashboard/helper";
 import EmployeeSelfCalender from "../../employeeProfile/dashboard/components/EmployeeSelfCalendar/index.jsx";
 import EmployeeSelfDashboardHeader from "../../employeeProfile/dashboard/components/EmployeeSelfDashboardHeader/index.jsx";
 import EmployeeSelfManagerList from "../../employeeProfile/dashboard/components/EmployeeSelfManagerList";
@@ -33,6 +36,10 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
   const [policyLanding, setPolicyLanding] = useState([]);
   const [show, setShow] = useState(false);
   const [singleNoticeData, setSingleNoticeData] = useState("");
+  const [filterIndex, setFilterIndex] = useState(0);
+  const [filterData, setFilterData] = useState([]);
+  const [rowDto, setRowDto] = useState([]);
+  const [loadingForBirth, setLoadingForBirth] = useState(false);
 
   useEffect(() => {
     setLoading(true);
@@ -46,6 +53,19 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
     // eslint-disable-next-line
     [employeeDashboard]
   );
+
+  const tabName = [
+    { name: "BirthDay", noLeftRadius: false, noRadius: false },
+    { name: "Work Anniversary", noLeftRadius: false, noRadius: true },
+  ];
+
+  const getData = () => {
+    getBirthAnniversary(setRowDto, setLoadingForBirth);
+  };
+
+  useEffect(() => {
+    getData();
+  }, []);
 
   return (
     <>
@@ -179,7 +199,7 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
         </div>
         <div
           className="row mx-0 w-100"
-          style={{ height: "350px", marginTop: "12px" }}
+          style={{ height: "500px", marginTop: "12px" }}
         >
           <div className="col-md-9 h-100 pl-0">
             <div className="row mx-0 w-100 h-100">
@@ -209,7 +229,7 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
                       My Applications
                     </h2>
                   </div>
-                  <div className="tableOne" style={{ height: "90%" }}>
+                  <div className="tableOne" style={{ height: "40%" }}>
                     {!employeeDashboard?.applicationPendingViewModels
                       ?.length ? (
                       <NoResult />
@@ -270,6 +290,204 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
                         </tbody>
                       </table>
                     )}
+                  </div>
+                  <div className="row mt-1">
+                    <div className="col-lg-6">
+                      <div
+                        style={{
+                          height: "50px",
+                          display: "grid",
+                          placeItems: "center",
+                        }}
+                      >
+                        <h2
+                          style={{
+                            width: "100%",
+                            fontSize: "1rem",
+                            color: gray500,
+                            fontWeight: 600,
+                            paddingLeft: "10px",
+                          }}
+                        >
+                          Birthday
+                        </h2>
+                      </div>
+                      <div style={{ height: "180px", overflow: "auto" }}>
+                        <div className="tableOne h-100">
+                          <table
+                            className=""
+                            style={{ borderBottom: "1px solid #F2F4F7" }}
+                          >
+                            <thead>
+                              <tr>
+                                <th style={{ borderTop: "none" }}>
+                                  <p
+                                    style={{
+                                      color: gray400,
+                                      paddingLeft: "8px",
+                                    }}
+                                  >
+                                    SL
+                                  </p>
+                                </th>
+                                <th style={{ borderTop: "none" }}>
+                                  <p
+                                    style={{
+                                      color: gray400,
+                                      textAlign: "center",
+                                    }}
+                                  >
+                                    Description
+                                  </p>
+                                </th>
+                              </tr>
+                            </thead>
+                            <tbody>
+                              {rowDto?.birthDayModel?.length > 0 ? (
+                                <>
+                                  {rowDto?.birthDayModel?.map((item, i) => (
+                                    <>
+                                      <tr key={i}>
+                                        <td
+                                          style={{
+                                            borderTop: "1px solid #F2F4F7",
+                                            position: "sticky",
+                                            left: 0,
+                                            zIndex: 1,
+                                            background: "#fff",
+                                          }}
+                                        >
+                                          <p
+                                            style={{
+                                              color: "gray700",
+                                              paddingLeft: "8px",
+                                            }}
+                                          >
+                                            {i + 1}
+                                          </p>
+                                        </td>
+                                        <td
+                                          style={{
+                                            borderTop: "1px solid #F2F4F7",
+                                          }}
+                                        >
+                                          <p
+                                            style={{
+                                              textAlign: "left",
+                                              color: gray700,
+                                            }}
+                                          >
+                                            {`${item?.birthDayMessage} on ${item?.birthDay}`}
+                                          </p>
+                                        </td>
+                                      </tr>
+                                    </>
+                                  ))}
+                                </>
+                              ) : (
+                                <p>Today There Is No BirthDay</p>
+                              )}
+                            </tbody>
+                          </table>
+                        </div>
+                      </div>
+                    </div>
+                    <div className="col-lg-6">
+                      <div
+                        style={{
+                          height: "50px",
+                          display: "grid",
+                          placeItems: "center",
+                        }}
+                      >
+                        <h2
+                          style={{
+                            width: "100%",
+                            fontSize: "1rem",
+                            color: gray500,
+                            fontWeight: 600,
+                            paddingLeft: "10px",
+                          }}
+                        >
+                          Work Annivarsay
+                        </h2>
+                      </div>
+                      <div style={{ height: "180px" }}>
+                        <div className="tableOne h-100">
+                          <table
+                            className=""
+                            style={{ borderBottom: "1px solid #F2F4F7" }}
+                          >
+                            <thead>
+                              <tr>
+                                <th style={{ borderTop: "none" }}>
+                                  <p
+                                    style={{
+                                      color: gray400,
+                                      paddingLeft: "8px",
+                                    }}
+                                  >
+                                    SL
+                                  </p>
+                                </th>
+                                <th style={{ borderTop: "none" }}>
+                                  <p
+                                    style={{
+                                      color: gray400,
+                                      textAlign: "center",
+                                    }}
+                                  >
+                                    Description
+                                  </p>
+                                </th>
+                              </tr>
+                            </thead>
+                            <tbody>
+                              {rowDto?.anniversaryModel?.length > 0 ? (
+                                <>
+                                  {rowDto?.anniversaryModel?.map((item, i) => (
+                                    <>
+                                      <tr key={i}>
+                                        <td
+                                          style={{
+                                            borderTop: "1px solid #F2F4F7",
+                                          }}
+                                        >
+                                          <p
+                                            style={{
+                                              color: gray700,
+                                              paddingLeft: "8px",
+                                            }}
+                                          >
+                                            {i + 1}
+                                          </p>
+                                        </td>
+                                        <td
+                                          style={{
+                                            borderTop: "1px solid #F2F4F7",
+                                          }}
+                                        >
+                                          <p
+                                            style={{
+                                              textAlign: "left",
+                                              color: gray700,
+                                            }}
+                                          >
+                                            {`${item?.anniversarMessage} on ${item?.anniversaryDay}`}
+                                          </p>
+                                        </td>
+                                      </tr>
+                                    </>
+                                  ))}
+                                </>
+                              ) : (
+                                <p>Today There Is No Work Annivarsay</p>
+                              )}
+                            </tbody>
+                          </table>
+                        </div>
+                      </div>
+                    </div>
                   </div>
                 </div>
               </div>
@@ -391,6 +609,9 @@ const SelfDashboardLanding = ({ setDashboardRoles, setLoading }) => {
             />
           </div>
         </div>
+        {/* <div className="row mx-0 w-100" style={{height:'250px', marginTop:'20px'}}>
+
+        </div> */}
       </div>
       <ViewModal
         size="lg"
