@@ -68,7 +68,7 @@ function Designation() {
   // for view Modal
   const handleViewClose = () => setViewModal(false);
 
-  const { orgId, buId } = useSelector(
+  const { orgId, buId, wId } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -86,8 +86,8 @@ function Designation() {
   }, []);
 
   useEffect(() => {
-    getAllDesignation(orgId, buId, setRowDto, setAllData, setLoading);
-  }, [orgId, buId]);
+    getAllDesignation(orgId, buId, setRowDto, setAllData, setLoading, wId);
+  }, [orgId, buId, wId]);
 
   const { permissionList } = useSelector((state) => state?.auth, shallowEqual);
 
