@@ -1,7 +1,7 @@
 import {
   AddOutlined,
   SearchOutlined,
-  SettingsBackupRestoreOutlined
+  SettingsBackupRestoreOutlined,
 } from "@mui/icons-material";
 
 import { Form, Formik } from "formik";
@@ -59,7 +59,7 @@ function LoanTypeCreate() {
     setId("");
   };
 
-  const { orgId, buId } = useSelector(
+  const { orgId, buId, wId } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -73,8 +73,8 @@ function LoanTypeCreate() {
   }, []);
 
   useEffect(() => {
-    getAllGlobalLoanType(setRowDto, setAllData, setLoading);
-  }, []);
+    getAllGlobalLoanType(wId, setRowDto, setAllData, setLoading);
+  }, [wId]);
 
   const { permissionList } = useSelector((state) => state?.auth, shallowEqual);
 
