@@ -59,7 +59,7 @@ const MgtExpenseApplicationCreate = () => {
   const params = useParams();
   const dispatch = useDispatch();
 
-  const { orgId, buId, employeeId, wgId } = useSelector(
+  const { orgId, buId, employeeId, wgId, wId } = useSelector(
     (state) => state?.auth?.profileData,
     shallowEqual
   );
@@ -122,7 +122,7 @@ const MgtExpenseApplicationCreate = () => {
 
   const getExpenseTypeDDL = (orgId) => {
     getExpenseType(
-      `/SaasMasterData/GetAllEmpExpenseType?IntAccountId=${orgId}`,
+      `/SaasMasterData/GetAllEmpExpenseType?IntAccountId=${orgId}&workplaceId=${wId}`,
       (data) => {
         const modifyData =
           data?.length > 0 &&
