@@ -201,7 +201,7 @@ export const iouLandingTableColumn = (page, paginationSize, history) => {
       dataIndex: "numIOUAmount",
       className: "text-right",
       sort: true,
-      render: (record) => numberWithCommas(record?.numIOUAmount),
+      // render: (record) => numberWithCommas(record?.numIOUAmount),
       filter: false,
       fieldType: "number",
     },
@@ -210,14 +210,14 @@ export const iouLandingTableColumn = (page, paginationSize, history) => {
       dataIndex: "numAdjustedAmount",
       className: "text-right",
       sort: true,
-      render: (record) => numberWithCommas(record?.numAdjustedAmount),
+      // render: (record) => numberWithCommas(record?.numAdjustedAmount),
       filter: false,
       fieldType: "number",
     },
     {
       title: "Adjusted by Accounts",
       dataIndex: "numReceivableAmount",
-      className: "text-right",
+      // className: "text-center",
       sort: true,
       render: (record) => numberWithCommas(record?.numReceivableAmount),
       filter: false,
@@ -226,17 +226,17 @@ export const iouLandingTableColumn = (page, paginationSize, history) => {
     },
     {
       title: "Pay to Accounts",
-      dataIndex: "numPendingAdjAmount",
-      className: "text-right",
+      dataIndex: "pendingAdjAmount",
+      // className: "text-right",
       sort: true,
-      render: (record) => numberWithCommas(record?.numPendingAdjAmount) || 0,
+      render: (record) => numberWithCommas(record?.pendingAdjAmount) || 0,
       filter: false,
       fieldType: "number",
     },
     {
       title: "Receive from Accounts",
       dataIndex: "numPayableAmount",
-      className: "text-right",
+      // className: "text-right",
       sort: true,
       render: (record) => numberWithCommas(record?.numPayableAmount),
       filter: false,
