@@ -101,7 +101,6 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
 
   const onRoleAdd = (values, setFieldValue) => {
     let modifyArr = [];
-
     // role Exist
     const isRoleExist = (values) => {
       let roleExist = false;
@@ -184,66 +183,15 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
       );
     };
 
-    // same work filed check
-    const isSameWorkField = (fieldValue, typeValue) => {
-      return rowDto.some(
-        (itm) =>
-          itm?.intOrganizationReffId === fieldValue &&
-          itm?.intOrganizationTypeId === typeValue
-      );
-    };
-
     // modify filter Arr
     const modifyFinalArr = (filedType, values) => {
-      if (filedType === "wing") {
-        return isSameWorkField(values?.orgName?.value, 2)
-          ? modifyArr.filter(
-            (itm) => itm?.intOrganizationReffId !== values?.orgName?.value
-          )
-          : modifyArr;
-      }
-
-      if (filedType === "soleDepo") {
-        return isSameWorkField(values?.wing?.value, 4)
-          ? modifyArr.filter(
-            (itm) =>
-              itm?.intOrganizationReffId !== values?.orgName?.value &&
-              itm?.intOrganizationReffId !== values?.wing?.value
-          )
-          : modifyArr;
-      }
-
-      if (filedType === "region") {
-        return isSameWorkField(values?.soleDepo?.value, 5)
-          ? modifyArr.filter(
-            (itm) =>
-              itm?.intOrganizationReffId !== values?.orgName?.value &&
-              itm?.intOrganizationReffId !== values?.wing?.value &&
-              itm?.intOrganizationReffId !== values?.soleDepo?.value
-          )
-          : modifyArr;
-      }
-
-      if (filedType === "area") {
-        return isSameWorkField(values?.region?.value, 6)
-          ? modifyArr.filter(
-            (itm) =>
-              itm?.intOrganizationReffId !== values?.orgName?.value &&
-              itm?.intOrganizationReffId !== values?.wing?.value &&
-              itm?.intOrganizationReffId !== values?.soleDepo?.value &&
-              itm?.intOrganizationReffId !== values?.region?.value
-          )
-          : modifyArr;
-      }
+      return;
     };
-
     // workplace Group
     if (isAllDataCheck(2)) {
       return toast.warn("Workplace Group has all data exsist...");
     }
-
     modifyArr = [
-      ...modifyArr,
       {
         intOrganizationTypeId: +values?.orgType?.value,
         strOrganizationTypeName: values?.orgType?.label,
@@ -252,148 +200,14 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
       },
     ];
 
-    // wing
-    if (isAllDataCheck(4)) {
-      return toast.warn("Wing has all data exsist...");
-    }
-
-    if (values?.wing?.value === 0 || values?.wing?.value) {
-      if (values?.wing?.label === "All") {
-        let modifyWing = rowDto.filter((itm) => itm?.intOrganizationTypeId < 4);
-        modifyArr = [
-          ...modifyWing,
-          {
-            intOrganizationTypeId: 4,
-            strOrganizationTypeName: "Wing",
-            intOrganizationReffId: values?.wing?.value,
-            strOrganizationReffName: values?.wing?.label,
-          },
-        ];
-
-        setRowDto([...modifyArr]);
-      } else {
-        modifyArr = [
-          ...modifyFinalArr("wing", values),
-          {
-            intOrganizationTypeId: 4,
-            strOrganizationTypeName: "Wing",
-            intOrganizationReffId: values?.wing?.value,
-            strOrganizationReffName: values?.wing?.label,
-          },
-        ];
-
-        setRowDto([...rowDto, ...modifyArr]);
-      }
-    }
-
-    // soleDepo
-    if (isAllDataCheck(5)) {
-      return toast.warn("Wing has all data exsist...");
-    }
-
-    if (values?.soleDepo?.value === 0 || values?.soleDepo?.value) {
-      if (values?.soleDepo?.label === "All") {
-        let modifySoleDepo = rowDto.filter(
-          (itm) => itm?.intOrganizationTypeId < 5
-        );
-        modifyArr = [
-          ...modifySoleDepo,
-          {
-            intOrganizationTypeId: 5,
-            strOrganizationTypeName: "Sole Depo",
-            intOrganizationReffId: values?.soleDepo?.value,
-            strOrganizationReffName: values?.soleDepo?.label,
-          },
-        ];
-
-        setRowDto([...modifyArr]);
-      } else {
-        modifyArr = [
-          ...modifyFinalArr("soleDepo", values),
-          {
-            intOrganizationTypeId: 5,
-            strOrganizationTypeName: "Sole Depo",
-            intOrganizationReffId: values?.soleDepo?.value,
-            strOrganizationReffName: values?.soleDepo?.label,
-          },
-        ];
-        setRowDto([...rowDto, ...modifyArr]);
-      }
-    }
-
-    // region
-    if (isAllDataCheck(6)) {
-      return toast.warn("Wing has all data exsist...");
-    }
-
-    if (values?.region?.value === 0 || values?.region?.value) {
-      if (values?.region?.label === "All") {
-        let modifyRegion = rowDto.filter(
-          (itm) => itm?.intOrganizationTypeId < 6
-        );
-        modifyArr = [
-          ...modifyRegion,
-          {
-            intOrganizationTypeId: 6,
-            strOrganizationTypeName: "Region",
-            intOrganizationReffId: values?.region?.value,
-            strOrganizationReffName: values?.region?.label,
-          },
-        ];
-
-        setRowDto([...modifyArr]);
-      } else {
-        modifyArr = [
-          ...modifyFinalArr("region", values),
-          {
-            intOrganizationTypeId: 6,
-            strOrganizationTypeName: "Region",
-            intOrganizationReffId: values?.region?.value,
-            strOrganizationReffName: values?.region?.label,
-          },
-        ];
-        setRowDto([...rowDto, ...modifyArr]);
-      }
-    }
+    setRowDto([...rowDto, ...modifyArr]);
 
     // region
     if (isAllDataCheck(7)) {
       return toast.warn("Wing has all data exsist...");
     }
-
-    if (values?.area?.value === 0 || values?.area?.value) {
-      if (values?.area?.label === "All") {
-        let modifyRegion = rowDto.filter(
-          (itm) => itm?.intOrganizationTypeId < 7
-        );
-        modifyArr = [
-          ...modifyRegion,
-          {
-            intOrganizationTypeId: 7,
-            strOrganizationTypeName: "Area",
-            intOrganizationReffId: values?.area?.value,
-            strOrganizationReffName: values?.area?.label,
-          },
-        ];
-
-        setRowDto([...modifyArr]);
-      } else {
-        modifyArr = [
-          ...modifyFinalArr("area", values),
-          {
-            intOrganizationTypeId: 7,
-            strOrganizationTypeName: "Area",
-            intOrganizationReffId: values?.area?.value,
-            strOrganizationReffName: values?.area?.label,
-          },
-        ];
-        setRowDto([...rowDto, ...modifyArr]);
-      }
-    }
-
     setFieldValue("orgType", "");
     setFieldValue("orgName", "");
-
     setFieldValue("wing", "");
     setFieldValue("soleDepo", "");
     setFieldValue("region", "");
@@ -475,61 +289,65 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
                   </div>
                   <div className="row">
                     <div className="col-md-3">
-                      <label className="mb-2">Organization Type</label>
-                      <FormikSelect
-                        classes="input-sm"
-                        styles={customStyles}
-                        name="businessUnit"
-                        options={organizationTypeList || []}
-                        value={values?.orgType}
-                        isDisabled={!values?.employeeName?.value}
-                        onChange={(valueOption) => {
-                          setFieldValue("orgName", "");
-                          setFieldValue("orgType", valueOption);
-                          setOrganizationDDLFunc(
-                            wgId,
-                            buId,
-                            employeeId,
-                            valueOption,
-                            setOrganizationDDL
-                          );
-                        }}
-                        errors={errors}
-                        touched={touched}
-                        placeholder=" "
-                      />
+                      <div className="input-field-main">
+                        <label>Organization Type</label>
+                        <FormikSelect
+                          classes="input-sm"
+                          styles={customStyles}
+                          name="businessUnit"
+                          options={organizationTypeList || []}
+                          value={values?.orgType}
+                          isDisabled={!values?.employeeName?.value}
+                          onChange={(valueOption) => {
+                            setFieldValue("orgName", "");
+                            setFieldValue("orgType", valueOption);
+                            setOrganizationDDLFunc(
+                              wgId,
+                              buId,
+                              employeeId,
+                              valueOption,
+                              setOrganizationDDL
+                            );
+                          }}
+                          errors={errors}
+                          touched={touched}
+                          placeholder=" "
+                        />
+                      </div>
                     </div>
                     <div className="col-md-3">
-                      <label className="mb-2">Organization Name</label>
-                      <FormikSelect
-                        isDisabled={!values?.orgType}
-                        classes="input-sm"
-                        styles={customStyles}
-                        name="orgName"
-                        options={organizationDDL || []}
-                        value={values?.orgName}
-                        onChange={(valueOption) => {
-                          setFieldValue("orgName", valueOption);
+                      <div className="input-field-main">
+                        <label>Organization Name</label>
+                        <FormikSelect
+                          isDisabled={!values?.orgType}
+                          classes="input-sm"
+                          styles={customStyles}
+                          name="orgName"
+                          options={organizationDDL || []}
+                          value={values?.orgName}
+                          onChange={(valueOption) => {
+                            setFieldValue("orgName", valueOption);
 
-                          if (valueOption?.label === "Marketing") {
-                            setFieldValue("wing", "");
-                            setFieldValue("soleDepo", "");
-                            setFieldValue("region", "");
-                            setFieldValue("area", "");
+                            if (valueOption?.label === "Marketing") {
+                              setFieldValue("wing", "");
+                              setFieldValue("soleDepo", "");
+                              setFieldValue("region", "");
+                              setFieldValue("area", "");
 
-                            getPeopleDeskWithoutAllDDL(
-                              `/PeopleDeskDDL/PeopleDeskAllDDL?DDLType=WingDDL&BusinessUnitId=${buId}&WorkplaceGroupId=${valueOption?.value}&ParentTerritoryId=0`,
-                              "WingId",
-                              "WingName",
-                              setWingDDL
-                            );
-                          }
-                        }}
-                        errors={errors}
-                        touched={touched}
-                        placeholder=" "
-                        isClearable={false}
-                      />
+                              getPeopleDeskWithoutAllDDL(
+                                `/PeopleDeskDDL/PeopleDeskAllDDL?DDLType=WingDDL&BusinessUnitId=${buId}&WorkplaceGroupId=${valueOption?.value}&ParentTerritoryId=0`,
+                                "WingId",
+                                "WingName",
+                                setWingDDL
+                              );
+                            }
+                          }}
+                          errors={errors}
+                          touched={touched}
+                          placeholder=" "
+                          isClearable={false}
+                        />
+                      </div>
                     </div>
 
                     {/* marketing setup */}
@@ -568,7 +386,7 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
                     )}
 
                     {/* wing */}
-                    {values?.wing && (
+                    {values?.wing && values?.wing?.value !== 0 && (
                       <div className="col-md-3">
                         <div className="input-field-main">
                           <label>Sole Depo</label>
@@ -601,7 +419,7 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
                     )}
 
                     {/* soleDepo */}
-                    {values?.soleDepo && (
+                    {values?.soleDepo && values?.soleDepo?.value !== 0 && (
                       <div className="col-md-3">
                         <div className="input-field-main">
                           <label>Region</label>
@@ -633,7 +451,7 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
                     )}
 
                     {/* region */}
-                    {values?.region && (
+                    {values?.region && values?.region?.value !== 0 && (
                       <div className="col-md-3">
                         <div className="input-field-main">
                           <label>Area</label>
@@ -665,7 +483,7 @@ const CreateRoleExtension = ({ setCreateOrUpdate }) => {
                           !values?.orgType
                         }
                         className="btn btn-green"
-                        style={{ marginTop: "27px" }}
+                        style={{ marginTop: "23px" }}
                         type="button"
                         onClick={() => {
                           onRoleAdd(values, setFieldValue);
