@@ -2,7 +2,6 @@ import { Form, Formik } from "formik";
 import React, { useEffect, useState } from "react";
 import { shallowEqual, useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
-import * as Yup from "yup";
 import moment from "moment";
 import MasterFilter from "../../../../common/MasterFilter";
 import NotPermittedPage from "../../../../common/notPermitted/NotPermittedPage";
@@ -12,7 +11,18 @@ import NoResult from "./../../../../common/NoResult";
 import AddEditFormComponent from "./addEditForm";
 import ResetButton from "./../../../../common/ResetButton";
 import "./calendar.css";
-import { columns } from "./helper";
+import {
+  bgColors,
+  colors,
+  columns,
+  getData,
+  handleChangePage,
+  handleChangeRowsPerPage,
+  initData,
+  initHeaderList,
+  statusDDL,
+  validationSchema,
+} from "./helper";
 import { Clear, SettingsBackupRestoreOutlined } from "@mui/icons-material";
 import Calender from "./component/Calender";
 import { IconButton, Popover } from "@mui/material";
@@ -20,50 +30,43 @@ import { gray900 } from "../../../../utility/customColor";
 import PeopleDeskTable, {
   paginationSize,
 } from "../../../../common/peopleDeskTable";
-import axios from "axios";
-import {
-  createPayloadStructure,
-  setHeaderListDataDynamically,
-} from "../../../../common/peopleDeskTable/helper";
+
 import FormikSelect from "../../../../common/FormikSelect";
 import { customStyles } from "../../../../utility/selectCustomStyle";
 
-const initData = {
-  searchString: "",
-  allSelected: false,
-  // master filter
-  workplace: "",
-  department: "",
-  designation: "",
-  supervisor: "",
-  employmentType: "",
-  employee: "",
-  assignStatus: { value: "all", label: "All" },
-  salaryStatus: "",
-};
-
-const statusDDL = [
-  { value: 0, label: "All" },
-  { value: 1, label: "Assigned" },
-  { value: 2, label: "Not Assigned" },
-];
-
-const validationSchema = Yup.object({});
-
 function Calendar() {
+  // redux
+  let permission = null;
+  const { permissionList } = useSelector((state) => state?.auth, shallowEqual);
+
+  const { orgId, buId, wgId, wgName, wId } = useSelector(
+    (state) => state?.auth?.profileData,
+    shallowEqual
+  );
+
+  permissionList.forEach((item) => {
+    if (item?.menuReferenceId === 44) {
+      permission = item;
+    }
+  });
+  const dispatch = useDispatch();
+
   // row Data
   const [rowDto, setRowDto] = useState([]);
   const [singleData, setSingleData] = useState([]);
   // const [checked, setChecked] = useState([]);
+
+  // shift info
   const [singleShiftData, setSingleShiftData] = useState([]);
   const [uniqueShift, setUniqueShift] = useState([]);
+
+  // colors
   const [uniqueShiftColor, setUniqueShiftColor] = useState({});
   const [uniqueShiftBg, setUniqueShiftBg] = useState({});
-  // modal
+
+  // modals
   const [createModal, setCreateModal] = useState(false);
   const handleCreateClose = () => setCreateModal(false);
-
-  // master filter
   const [anchorEl, setAnchorEl] = React.useState(null);
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -71,50 +74,17 @@ function Calendar() {
   const open = Boolean(anchorEl);
   const id = open ? "simple-popover" : undefined;
 
-  const { orgId, buId, wgId, wgName, wId } = useSelector(
-    (state) => state?.auth?.profileData,
-    shallowEqual
-  );
-  const colors = [
-    "#299647",
-    "#B54708",
-    "#B42318",
-    "#6927DA",
-    "#3538CD",
-    "#667085",
-    "#667085",
-  ];
-  // eslint-disable-next-line
-  const [bgColors, setBgColors] = useState([
-    "#E6F9E9",
-    "#FEF0C7",
-    "#FEE4E2",
-    "#ECE9FE",
-    "#E0EAFF",
-    "#F2F4F7",
-    "#FEF0D7",
-  ]);
-
   const [anchorEl2, setAnchorEl2] = useState(null);
   const open2 = Boolean(anchorEl2);
   const id2 = open2 ? "simple-popover" : undefined;
+
+  // pagination
   const [pages, setPages] = useState({
     current: 1,
     pageSize: paginationSize,
     total: 0,
   });
-
-  const initHeaderList = {
-    designationList: [],
-    departmentList: [],
-    supervisorNameList: [],
-    wingNameList: [],
-    soleDepoNameList: [],
-    regionNameList: [],
-    areaNameList: [],
-    territoryNameList: [],
-    employmentTypeList: [],
-  };
+  // peopledesk states
   const [landingLoading, setLandingLoading] = useState(false);
   const [filterOrderList, setFilterOrderList] = useState([]);
   const [initialHeaderListData, setInitialHeaderListData] = useState({});
@@ -126,154 +96,33 @@ function Calendar() {
   const [empIDString, setEmpIDString] = useState("");
   const [isAssignAll, setIsAssignAll] = useState(false);
 
-  // landing api call
-  const getDataApiCall = async (
-    modifiedPayload,
-    pagination,
-    searchText,
-    checkedList,
-    currentFilterSelection,
-    checkedHeaderList,
-    isAssigned = null
-  ) => {
-    setLandingLoading(true);
-    try {
-      const payload = {
-        businessUnitId: buId,
-        workplaceGroupId: wgId,
-        isNotAssign: isAssigned === 1 ? false : isAssigned === 2 ? true : null,
-        workplaceId: wId,
-        pageNo: pagination.current,
-        pageSize: pagination.pageSize,
-        isPaginated: true,
-        isHeaderNeed: true,
-        searchTxt: searchText || "",
-      };
-
-      const res = await axios.post(`/Employee/CalendarAssignFilter`, {
-        ...payload,
-        ...modifiedPayload,
-      });
-      if (res?.data?.data) {
-        setLandingLoading(true);
-        setHeaderListDataDynamically({
-          currentFilterSelection,
-          checkedHeaderList,
-          headerListKey: "calendarAssignHeader",
-          headerList,
-          setHeaderList,
-          response: res?.data,
-          filterOrderList,
-          setFilterOrderList,
-          initialHeaderListData,
-          setInitialHeaderListData,
-          // setEmpLanding,
-          setPages,
-        });
-
-        setEmpIDString(res?.data?.employeeIdList);
-        const modifiedData = res?.data?.data?.map((item, index) => ({
-          ...item,
-          initialSerialNumber: index + 1,
-          isSelected: checkedList?.find(
-            ({ employeeCode }) => item?.employeeCode === employeeCode
-          )
-            ? true
-            : false,
-        }));
-
-        setRowDto(modifiedData);
-        setLandingLoading(false);
-      } else {
-        setRowDto([]);
-      }
-      setLandingLoading(false);
-    } catch (error) {
-      setLandingLoading(false);
-    }
-  };
-
-  const getData = async (
-    pagination,
-    searchText = "",
-    checkedList = [],
-    currentFilterSelection = -1,
-    filterOrderList = [],
-    checkedHeaderList = { ...initHeaderList },
-    isAssigned
-  ) => {
-    const modifiedPayload = createPayloadStructure({
-      initHeaderList,
-      currentFilterSelection,
-      checkedHeaderList,
-      filterOrderList,
-    });
-
-    getDataApiCall(
-      modifiedPayload,
-      pagination,
-      searchText,
-      checkedList,
-      currentFilterSelection,
-      checkedHeaderList,
-      isAssigned
-    );
-  };
-  const dispatch = useDispatch();
+  // sidebar
   useEffect(() => {
     dispatch(setFirstLevelNameAction("Administration"));
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  // landing data
   useEffect(() => {
-    getData(pages);
+    getData(
+      { current: 1, pageSize: paginationSize, total: 0 },
+      setLandingLoading,
+      buId,
+      wgId,
+      wId,
+      headerList,
+      setHeaderList,
+      setFilterOrderList,
+      initialHeaderListData,
+      setInitialHeaderListData,
+      setPages,
+      setEmpIDString,
+      setRowDto
+    );
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [buId, wgId, wId]);
 
-  const handleChangePage = (_, newPage, searchText) => {
-    setPages((prev) => {
-      return { ...prev, current: newPage };
-    });
-    getData(
-      {
-        current: newPage,
-        pageSize: pages?.pageSize,
-        total: pages?.total,
-      },
-      searchText,
-      checkedList,
-      -1,
-      filterOrderList,
-      checkedHeaderList
-    );
-  };
-
-  const handleChangeRowsPerPage = (event, searchText) => {
-    setPages((prev) => {
-      return { current: 1, total: pages?.total, pageSize: +event.target.value };
-    });
-    getData(
-      {
-        current: 1,
-        pageSize: +event.target.value,
-        total: pages?.total,
-      },
-      searchText,
-      checkedList,
-      -1,
-      filterOrderList,
-      checkedHeaderList
-    );
-  };
-  const { permissionList } = useSelector((state) => state?.auth, shallowEqual);
-
-  let permission = null;
-  permissionList.forEach((item) => {
-    if (item?.menuReferenceId === 44) {
-      permission = item;
-    }
-  });
-
+  // assign colors to shift on shift load
   useEffect(() => {
     setUniqueShift([]);
     if (singleShiftData?.length > 0) {
@@ -342,12 +191,25 @@ function Calendar() {
                                 onClick={() => {
                                   getData(
                                     { current: 1, pageSize: paginationSize },
+                                    setLandingLoading,
+                                    buId,
+                                    wgId,
+                                    wId,
+                                    headerList,
+                                    setHeaderList,
+                                    setFilterOrderList,
+                                    initialHeaderListData,
+                                    setInitialHeaderListData,
+                                    setPages,
+                                    setEmpIDString,
+                                    setRowDto,
                                     "",
                                     [],
                                     -1,
                                     filterOrderList,
                                     checkedHeaderList
                                   );
+
                                   // setRowDto(allData);
                                   setCheckedList([]);
                                   setFieldValue("searchString", "");
@@ -415,6 +277,18 @@ function Calendar() {
                                 setFieldValue("searchString", "");
                                 getData(
                                   { current: 1, pageSize: paginationSize },
+                                  setLandingLoading,
+                                  buId,
+                                  wgId,
+                                  wId,
+                                  headerList,
+                                  setHeaderList,
+                                  setFilterOrderList,
+                                  initialHeaderListData,
+                                  setInitialHeaderListData,
+                                  setPages,
+                                  setEmpIDString,
+                                  setRowDto,
                                   "",
                                   checkedList,
                                   -1,
@@ -436,6 +310,18 @@ function Calendar() {
                                 if (value) {
                                   getData(
                                     { current: 1, pageSize: paginationSize },
+                                    setLandingLoading,
+                                    buId,
+                                    wgId,
+                                    wId,
+                                    headerList,
+                                    setHeaderList,
+                                    setFilterOrderList,
+                                    initialHeaderListData,
+                                    setInitialHeaderListData,
+                                    setPages,
+                                    setEmpIDString,
+                                    setRowDto,
                                     value,
                                     checkedList,
                                     -1,
@@ -446,6 +332,18 @@ function Calendar() {
                                 } else {
                                   getData(
                                     { current: 1, pageSize: paginationSize },
+                                    setLandingLoading,
+                                    buId,
+                                    wgId,
+                                    wId,
+                                    headerList,
+                                    setHeaderList,
+                                    setFilterOrderList,
+                                    initialHeaderListData,
+                                    setInitialHeaderListData,
+                                    setPages,
+                                    setEmpIDString,
+                                    setRowDto,
                                     "",
                                     [],
                                     -1,
@@ -459,6 +357,18 @@ function Calendar() {
                                 setFieldValue("searchString", "");
                                 getData(
                                   { current: 1, pageSize: paginationSize },
+                                  setLandingLoading,
+                                  buId,
+                                  wgId,
+                                  wId,
+                                  headerList,
+                                  setHeaderList,
+                                  setFilterOrderList,
+                                  initialHeaderListData,
+                                  setInitialHeaderListData,
+                                  setPages,
+                                  setEmpIDString,
+                                  setRowDto,
                                   "",
                                   [],
                                   -1,
@@ -499,10 +409,49 @@ function Calendar() {
                         checkedHeaderList={checkedHeaderList}
                         setCheckedHeaderList={setCheckedHeaderList}
                         handleChangePage={(e, newPage) =>
-                          handleChangePage(e, newPage, values?.search)
+                          handleChangePage(
+                            e,
+                            newPage,
+                            values?.search,
+                            setLandingLoading,
+                            buId,
+                            wgId,
+                            wId,
+                            headerList,
+                            setHeaderList,
+                            setFilterOrderList,
+                            initialHeaderListData,
+                            setInitialHeaderListData,
+                            setPages,
+                            setEmpIDString,
+                            setRowDto,
+                            checkedList,
+                            pages,
+                            filterOrderList,
+                            checkedHeaderList
+                          )
                         }
                         handleChangeRowsPerPage={(e) =>
-                          handleChangeRowsPerPage(e, values?.search)
+                          handleChangeRowsPerPage(
+                            e,
+                            values?.search,
+                            setLandingLoading,
+                            buId,
+                            wgId,
+                            wId,
+                            headerList,
+                            setHeaderList,
+                            setFilterOrderList,
+                            initialHeaderListData,
+                            setInitialHeaderListData,
+                            setPages,
+                            setEmpIDString,
+                            setRowDto,
+                            checkedList,
+                            pages,
+                            filterOrderList,
+                            checkedHeaderList
+                          )
                         }
                         filterOrderList={filterOrderList}
                         setFilterOrderList={setFilterOrderList}
@@ -518,6 +467,18 @@ function Calendar() {
                               pageSize: paginationSize,
                               total: 0,
                             },
+                            setLandingLoading,
+                            buId,
+                            wgId,
+                            wId,
+                            headerList,
+                            setHeaderList,
+                            setFilterOrderList,
+                            initialHeaderListData,
+                            setInitialHeaderListData,
+                            setPages,
+                            setEmpIDString,
+                            setRowDto,
                             "",
                             [],
                             currentFilterSelection,
@@ -556,6 +517,18 @@ function Calendar() {
                 getData={(checked) =>
                   getData(
                     { current: 1, pageSize: paginationSize },
+                    setLandingLoading,
+                    buId,
+                    wgId,
+                    wId,
+                    headerList,
+                    setHeaderList,
+                    setFilterOrderList,
+                    initialHeaderListData,
+                    setInitialHeaderListData,
+                    setPages,
+                    setEmpIDString,
+                    setRowDto,
                     "",
                     checked,
                     -1,
@@ -572,6 +545,7 @@ function Calendar() {
                 rowDto={rowDto}
               />
 
+              {/* i button calendar view */}
               {singleShiftData.length > 0 ? (
                 <Popover
                   sx={{
