@@ -1,8 +1,8 @@
 import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
-import { persistStore, persistReducer, createTransform } from "redux-persist";
-import { rootReducer } from "./rootReducer";
-import storage from "redux-persist/lib/storage";
 import CryptoJS from "crypto-js";
+import { createTransform, persistReducer, persistStore } from "redux-persist";
+import storage from "redux-persist/lib/storage";
+import { rootReducer } from "./rootReducer";
 
 const encrypt = createTransform(
   (inboundState, key) => {
